requirements.txt
Django==1.7.4
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
Sphinx==1.2.3
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
django-extensions==1.5.0
-e git+git@github.com:BryanAke/django-rest-framework-mongoengine.git@3c131f9e56f767fcaee3bdc1e43a298014f57b1a#egg=django_rest_framework_mongoengine-origin/upstream_master
djangorestframework==3.2.0
docutils==0.12
drf-extensions==0.2.7
ipdb==0.8
ipython==3.2.1
mongoengine==0.8.7
nose==1.3.4
pymongo==2.8
pyreadline==2.0
pyzmq==14.5.0
six==1.9.0
tornado==4.0.2
